% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcol.R
\name{adcol}
\alias{adcol}
\title{Function to add a new column to the matrix}
\usage{
adcol(x, y, z, run)
}
\arguments{
\item{x}{levels vector of the new factor}

\item{y}{number of levels of the new factor}

\item{z}{level vector of the initial matrix}

\item{run}{initial run matrix}
}
\value{
matrix with the new run order
}
\description{
Function to add a new column to the matrix
}
\examples{
x = matrix(c(-1, 1), ncol = 1)
y = length(x)
z = c(2,2,2)
run=matrix(c(1,-1,1,-1,1,1,-1,-1), ncol=2)
adcol(x,y,z,run)
}
