% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runorder.R
\name{runorder}
\alias{runorder}
\title{Assignment-Expansion method}
\usage{
runorder(z)
}
\arguments{
\item{z}{vector with the levels of the factor}
}
\value{
order of experimentation with bias and number of level changes in balance
}
\description{
Assignment-Expansion method
}
\examples{
z<-c(2,2,2,2,2,2)
runorder(z)
z<-c(4,3,2,3,2)
runorder(z)
z<-c(3,3,2,4)
runorder(z)

}
