% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Función para sumar}
\usage{
suma(a, b)
}
\arguments{
\item{a}{primer número}

\item{b}{segundo número}
}
\value{
resultado de la suma
}
\description{
Función para sumar
}
\examples{
suma(a=4,b=6)
suma(a=7,b=3)
}
